+++ tests/testthat/test-textmodel.R
@@ -1,6 +1,7 @@
 require(quanteda)
 
-toks_test <- tokens(data_corpus_inaugural, remove_punct = TRUE)
+corp_test <- head(data_corpus_inaugural, 60)
+toks_test <- tokens(corp_test, remove_punct = TRUE)
 dfmt_test <- dfm(toks_test) %>%
     dfm_remove(stopwords("en"))
 toks_dict_test <- tokens_lookup(toks_test, data_dictionary_newsmap_en, level = 3)
@@ -14,13 +15,20 @@ test_that("textmodel_newsmap() works with different inputs", {
 
     smat <- xtabs( ~ docid(dfmt) + dfmt$Party, sparse = TRUE)
     map1 <- textmodel_newsmap(dfmt, smat)
+    expect_identical(map1$data, dfmt)
     expect_equal(names(coef(map1)), levels(dfmt$Party))
     expect_null(map1$weight)
 
     mat <- as.matrix(smat)
     map2 <- textmodel_newsmap(dfmt, mat)
+    expect_identical(map2$data, dfmt)
     expect_equal(names(coef(map2)), levels(dfmt$Party))
     expect_null(map2$weight)
+
+    map3 <- textmodel_newsmap(dfmt, mat, boolean = TRUE)
+    expect_identical(map3$data, dfmt)
+    expect_equal(names(coef(map3)), levels(dfmt$Party))
+    expect_false(identical(coef(map1), coef(map3)))
 
     expect_error(textmodel_newsmap(list(), smat))
     expect_error(textmodel_newsmap(dfmt, list()))
@@ -51,6 +59,22 @@ test_that("textmodel_newsmap() works with different inputs", {
     expect_false(all(map_loc$weight[1,] == map_avg$weight[1,]))
     expect_false(all(map_avg$weight[1,] == map_glo$weight[1,]))
 
+    expect_error(
+        textmodel_newsmap(dfmt, smat, smooth = -1),
+        "The value of smooth must be between 0 and Inf"
+    )
+    expect_error(
+        textmodel_newsmap(dfmt, smat, smooth = c(1, 2)),
+        "The length of smooth must be 1"
+    )
+    expect_error(
+        textmodel_newsmap(dfmt, smat, boolean = "yes"),
+        "The value of boolean cannot be NA"
+    )
+    expect_error(
+        textmodel_newsmap(dfmt, smat, drop_label = "no"),
+        "The value of drop_label cannot be NA"
+    )
 })
 
 
@@ -212,11 +236,11 @@ test_that("coef() and dictionary() are working", {
     expect_error(coef(map, select = "xx"),
                  "Selected class must be in the model")
     expect_error(coef(map, select = character()),
-                 "The length of select must be between 1 and 16")
+                 "The length of select must be between 1 and 17")
 
     # TODO: remove as.list()
     lis1 <- as.list(map)
-    expect_equal(length(lis1), 16)
+    expect_equal(length(lis1), 17)
     expect_true(all(sapply(lis1, is.character)))
     expect_true(all(lengths(as.list(lis1)) == 10))