+++ NEWS.md
@@ -1,3 +1,11 @@
+## Changes in v0.9.2
+
+* Fix tests for **quanteda** v4.3.1.
+
+## Changes in v0.9.1
+
+* Add `boolean` to `textmodel_newsmap()`.
+
 ## Changes in v0.9.0
 
 * Add Turkish seed dictionary